Backend/requirements-hf.txt
# Hugging Face Spaces - Production requirements for AstroPixel
# Optimized for free tier deployment

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# File handling
aiofiles==23.2.1
pillow==10.2.0

# Data processing
numpy==1.26.3
boto3==1.34.0
psutil==5.9.8

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
